Fix incorrect item getting edited
When attempting to edit a todo item, it was possible for the wrong item
to be selected to set the initial value. Sometimes, it would attempt to
select an initial value that didn't exist, causing the edit dialog to
simply never pop up. This was because the ID of the todo item was being
used as the index to select from an array of todo items.

Fixes #25